amd-vi: fix up r326152, the new width requires a wider type
This is my brain-o from extending the width at the last moment.


git-svn-id: svn+ssh://svn.freebsd.org/base/head@326153 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f

+++ sys/amd64/vmm/amd/amdvi_priv.h
@@ -230,7 +230,7 @@ struct amdvi_ctrl {
 	uint64_t :45;
 	uint32_t evt_tail:19;
 	uint64_t :45;
-	uint8_t	 status:19;
+	uint32_t status:19;
 	uint64_t :45;
 	uint64_t pad2;
 	uint8_t  :4;
